backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==12.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
Pillow==10.4.0
scipy==1.13.1
aiosqlite==0.20.0
onnxruntime-qnn==1.19.0
onnxruntime-genai==0.4.0
qai-hub>=0.15.0
qai-hub-models>=0.12.0
